# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PosLoginInput {
  /** @pattern ^\d{4}$ */
  pin: string;
}

/**
 * @nullable
 */
export type PosSessionRole = typeof PosSessionRole[keyof typeof PosSessionRole] | null;


export const PosSessionRole = {
  admin: 'admin',
  worker: 'worker',
} as const;

export interface PosSession {
  authenticated: boolean;
  /** @nullable */
  role: PosSessionRole;
}

export interface MenuItem {
  id: string;
  name: string;
  /** @minimum 0 */
  price: number;
  /**
     * @minimum 24
     * @maximum 27
     */
  tobaccoGrams: number;
  enabled: boolean;
}

export interface WorkerPermissions {
  menuItemIds: string[];
  seeRevenue: boolean;
  useDiscounts: boolean;
  addInventory: boolean;
  editPastInventory: boolean;
  editPastShifts: boolean;
}

export interface InventoryItem {
  id: string;
  name: string;
  category: string;
  brand: string;
  /** @minimum 0 */
  packageGrams: number;
  /** @minimum 0 */
  stock: number;
  unit: string;
}

export interface Container {
  id: string;
  name: string;
  /** @minimum 0 */
  tareGrams: number;
}

export type ShiftLineDiscount = typeof ShiftLineDiscount[keyof typeof ShiftLineDiscount];


export const ShiftLineDiscount = {
  NUMBER_0: 0,
  NUMBER_50: 50,
  NUMBER_100: 100,
} as const;

export interface ShiftLine {
  id: string;
  menuItemId: string;
  menuItemName: string;
  /** @minimum 0 */
  unitPrice: number;
  /**
     * @minimum 24
     * @maximum 27
     */
  tobaccoGrams: number;
  discount: ShiftLineDiscount;
  reason: string;
}

export interface Shift {
  id: string;
  date: string;
  master: string;
  lines: ShiftLine[];
  /** @minimum 0 */
  refills: number;
  /** @minimum 0 */
  helpersPay: number;
  comment: string;
  /**
     * @minimum 0
     * @nullable
     */
  fixedSalary: number | null;
  createdAt: string;
}

export interface PosState {
  menu: MenuItem[];
  permissions: WorkerPermissions;
  inventory: InventoryItem[];
  containers: Container[];
  shifts: Shift[];
  /** @minimum 0 */
  coalOpeningKg: number;
  updatedAt: string;
}