# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: pos
    description: Hookah lounge point of sale
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /pos/session:
    get:
      operationId: getPosSession
      tags: [pos]
      summary: Get the current POS session
      responses:
        "200":
          description: Current role, or no active session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PosSession"
    delete:
      operationId: logoutPos
      tags: [pos]
      summary: End the current POS session
      responses:
        "200":
          description: Session ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PosSession"
  /pos/login:
    post:
      operationId: loginPos
      tags: [pos]
      summary: Sign in with a four-digit POS PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PosLoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PosSession"
        "401":
          description: Incorrect PIN
  /pos/state:
    get:
      operationId: getPosState
      tags: [pos]
      summary: Load settings, menu, inventory, and shift history
      responses:
        "200":
          description: POS application state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PosState"
    put:
      operationId: savePosState
      tags: [pos]
      summary: Save POS state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PosState"
      responses:
        "200":
          description: Saved POS application state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PosState"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    PosLoginInput:
      type: object
      properties:
        pin:
          type: string
          pattern: "^\\d{4}$"
      required: [pin]
    PosSession:
      type: object
      properties:
        authenticated:
          type: boolean
        role:
          type: ["string", "null"]
          enum: [admin, worker, null]
      required: [authenticated, role]
    MenuItem:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        price: { type: number, minimum: 0 }
        tobaccoGrams: { type: integer, minimum: 24, maximum: 27 }
        enabled: { type: boolean }
      required: [id, name, price, tobaccoGrams, enabled]
    WorkerPermissions:
      type: object
      properties:
        menuItemIds:
          type: array
          items: { type: string }
        seeRevenue: { type: boolean }
        useDiscounts: { type: boolean }
        addInventory: { type: boolean }
        editPastInventory: { type: boolean }
        editPastShifts: { type: boolean }
      required: [menuItemIds, seeRevenue, useDiscounts, addInventory, editPastInventory, editPastShifts]
    InventoryItem:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        brand: { type: string }
        packageGrams: { type: number, minimum: 0 }
        stock: { type: number, minimum: 0 }
        unit: { type: string }
      required: [id, name, category, brand, packageGrams, stock, unit]
    Container:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        tareGrams: { type: number, minimum: 0 }
      required: [id, name, tareGrams]
    ShiftLine:
      type: object
      properties:
        id: { type: string }
        menuItemId: { type: string }
        menuItemName: { type: string }
        unitPrice: { type: number, minimum: 0 }
        tobaccoGrams: { type: integer, minimum: 24, maximum: 27 }
        discount: { type: number, enum: [0, 50, 100] }
        reason: { type: string }
      required: [id, menuItemId, menuItemName, unitPrice, tobaccoGrams, discount, reason]
    Shift:
      type: object
      properties:
        id: { type: string }
        date: { type: string }
        master: { type: string }
        lines:
          type: array
          items: { $ref: "#/components/schemas/ShiftLine" }
        refills: { type: integer, minimum: 0 }
        helpersPay: { type: number, minimum: 0 }
        comment: { type: string }
        fixedSalary: { type: ["number", "null"], minimum: 0 }
        createdAt: { type: string }
      required: [id, date, master, lines, refills, helpersPay, comment, fixedSalary, createdAt]
    PosState:
      type: object
      properties:
        menu:
          type: array
          items: { $ref: "#/components/schemas/MenuItem" }
        permissions:
          $ref: "#/components/schemas/WorkerPermissions"
        inventory:
          type: array
          items: { $ref: "#/components/schemas/InventoryItem" }
        containers:
          type: array
          items: { $ref: "#/components/schemas/Container" }
        shifts:
          type: array
          items: { $ref: "#/components/schemas/Shift" }
        coalOpeningKg: { type: number, minimum: 0 }
        updatedAt: { type: string }
      required: [menu, permissions, inventory, containers, shifts, coalOpeningKg, updatedAt]