# lib/api-zod/src/generated/types/container.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Container {
  id: string;
  name: string;
  /** @minimum 0 */
  tareGrams: number;
}