# lib/api-zod/src/generated/types/inventoryItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItem {
  id: string;
  name: string;
  category: string;
  brand: string;
  /** @minimum 0 */
  packageGrams: number;
  /** @minimum 0 */
  stock: number;
  unit: string;
}