# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: string;
  name: string;
  /** @minimum 0 */
  price: number;
  /**
     * @minimum 24
     * @maximum 27
     */
  tobaccoGrams: number;
  enabled: boolean;
}