# lib/api-zod/src/generated/types/posSession.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PosSessionRole } from './posSessionRole';

export interface PosSession {
  authenticated: boolean;
  /** @nullable */
  role: PosSessionRole;
}