# lib/api-zod/src/generated/types/posState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Container } from './container';
import type { InventoryItem } from './inventoryItem';
import type { MenuItem } from './menuItem';
import type { Shift } from './shift';
import type { WorkerPermissions } from './workerPermissions';

export interface PosState {
  menu: MenuItem[];
  permissions: WorkerPermissions;
  inventory: InventoryItem[];
  containers: Container[];
  shifts: Shift[];
  /** @minimum 0 */
  coalOpeningKg: number;
  updatedAt: string;
}