# lib/api-zod/src/generated/types/shift.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShiftLine } from './shiftLine';

export interface Shift {
  id: string;
  date: string;
  master: string;
  lines: ShiftLine[];
  /** @minimum 0 */
  refills: number;
  /** @minimum 0 */
  helpersPay: number;
  comment: string;
  /**
     * @minimum 0
     * @nullable
     */
  fixedSalary: number | null;
  createdAt: string;
}