# lib/api-zod/src/generated/types/shiftLine.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ShiftLineDiscount } from './shiftLineDiscount';

export interface ShiftLine {
  id: string;
  menuItemId: string;
  menuItemName: string;
  /** @minimum 0 */
  unitPrice: number;
  /**
     * @minimum 24
     * @maximum 27
     */
  tobaccoGrams: number;
  discount: ShiftLineDiscount;
  reason: string;
}