# lib/api-zod/src/generated/types/workerPermissions.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkerPermissions {
  menuItemIds: string[];
  seeRevenue: boolean;
  useDiscounts: boolean;
  addInventory: boolean;
  editPastInventory: boolean;
  editPastShifts: boolean;
}